requirements.txt
dune-client
pandas
matplotlib
requests
python-dotenv
